requirements.txt
streamlit==1.33.0
requests==2.31.0
PyMuPDF==1.24.1
Pillow==10.3.0
torch==2.2.2
torchvision==0.17.2
scikit-learn==1.4.2
numpy==1.26.4
transformers==4.39.3
fpdf2==2.7.9
fastapi==0.110.1
uvicorn==0.29.0
python-multipart==0.0.9
pinecone-client==3.2.2
google-generativeai==0.4.1
python-dotenv==1.0.1
beautifulsoup4==4.14.3